add transport timeout options (bug #72) style

+++ src/lib/wsman-xml-serialize.h
@@ -97,11 +97,11 @@ typedef struct __NameAliase NameAliase;
 
 //int do_serialize_uint(struct __XmlSerializationData* data, int valSize);
 
-void* ws_serializer_alloc(WsContextH cntx, size_t size);
+void* ws_serializer_alloc(WsContextH cntx, int size);
 //int do_serializer_free(WsContextH cntx, void* ptr);
 int ws_serializer_free(WsContextH cntx, void* ptr);
 
-void *xml_serializer_alloc(XmlSerializationData *data, size_t size, int zeroInit);
+void *xml_serializer_alloc(XmlSerializationData *data, int size, int zeroInit);
 
 int xml_serializer_free(XmlSerializationData *data, void *buf);
